+++ Routing/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+﻿using System.Runtime.CompilerServices;
+
+#if DEBUG
+[assembly: InternalsVisibleTo("Messerli.Routing.Test")]
+#endif